Add MergeSortMulti function
The `MergeSortMulti([][]int) [][]int` function takes a slice of lists and sorts each list within the slice.

This version of the function doesn't use synchronisation or channels to avoid data races. Instead it relies on the fact that each go routine updates a single element in the results slice.

Tracing has been enabled.